requirements.txt
discord-py
aiohttp
psycopg2-binary
sqlalchemy
python-dotenv
flask